edit chat endpoint

+++ models/chat.go
@@ -1,8 +1,8 @@
 package models
 
 type Chat struct {
-	Id        int `json:"id"`
-	UserId    int `json:"user_id"`
-	CreatedAt int `json:"created_at"`
-	DeletedAt int `json:"deleted_at"`
+	Id        int   `json:"id"`
+	UserId    int   `json:"user_id"`
+	CreatedAt int64 `json:"created_at"`
+	DeletedAt int64 `json:"deleted_at"`
 }
