+++ models/user.go
@@ -1,6 +1,6 @@
 package models
 
 type User struct {
-	Id           int `json:"id"`
-	RegisteredAt int `json:"registered_at"`
+	Id           int   `json:"id"`
+	RegisteredAt int64 `json:"registered_at"`
 }
